+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Change Log
+
+## v1.0.0
+- Getters and setters generation script
+- Constructor generation script
+
+## v1.1.0
+- Bug fixes: End If on constructor, breaking lines
+- Attributes list generation script